src/auth_service/requirements.txt
redis
bcrypt
pyjwt
python-dotenv
cryptography
qrcode[pil]
